Add files via upload
The following project will utilize a survey of Data Mining algorithms to evaluate the current strategies for Market-Basket analysis.  The purpose of Market-basket analysis is to increase our understanding of purchasing behaviors, and better predict which products are most likely to be purchased together.  Current strategies for Market -Basket analysis include the Apriori Algorithm and FP-growth algorithm. CSV file provided is used to apply data mining algorithms (apriori and FP-growth algorithm).